+++ libpod/networking_freebsd.go
@@ -15,9 +15,6 @@ import (
 
 	"github.com/containers/buildah/pkg/jail"
 	"github.com/containers/common/libnetwork/types"
-	"github.com/containers/podman/v4/libpod/define"
-	"github.com/containers/podman/v4/libpod/events"
-	"github.com/containers/podman/v4/pkg/util"
 	"github.com/containers/storage/pkg/lockfile"
 	"github.com/sirupsen/logrus"
 )
@@ -263,232 +260,8 @@ func (c *Container) inspectJoinedNetworkNS(networkns string) (q types.StatusBloc
 
 }
 
-// NetworkDisconnect removes a container from the network
-func (c *Container) NetworkDisconnect(nameOrID, netName string, force bool) error {
-	// only the bridge mode supports cni networks
-	if err := isBridgeNetMode(c.config.NetMode); err != nil {
-		return err
-	}
-
-	c.lock.Lock()
-	defer c.lock.Unlock()
-
-	networks, err := c.networks()
-	if err != nil {
-		return err
-	}
-
-	// check if network exists and if the input is a ID we get the name
-	// CNI only uses names so it is important that we only use the name
-	netName, err = c.runtime.normalizeNetworkName(netName)
-	if err != nil {
-		return err
-	}
-
-	_, nameExists := networks[netName]
-	if !nameExists && len(networks) > 0 {
-		return fmt.Errorf("container %s is not connected to network %s", nameOrID, netName)
-	}
-
-	if err := c.syncContainer(); err != nil {
-		return err
-	}
-	// get network status before we disconnect
-	networkStatus := c.getNetworkStatus()
-
-	if err := c.runtime.state.NetworkDisconnect(c, netName); err != nil {
-		return err
-	}
-
-	c.newNetworkEvent(events.NetworkDisconnect, netName)
-	if !c.ensureState(define.ContainerStateRunning, define.ContainerStateCreated) {
-		return nil
-	}
-
-	opts := types.NetworkOptions{
-		ContainerID:   c.config.ID,
-		ContainerName: getCNIPodName(c),
-	}
-	opts.PortMappings = c.convertPortMappings()
-	opts.Networks = map[string]types.PerNetworkOptions{
-		netName: networks[netName],
-	}
-
-	// update network status if container is running
-	oldStatus, statusExist := networkStatus[netName]
-	delete(networkStatus, netName)
-	c.state.NetworkStatus = networkStatus
-	err = c.save()
-	if err != nil {
-		return err
-	}
-
-	// Update resolv.conf if required
-	if statusExist {
-		stringIPs := make([]string, 0, len(oldStatus.DNSServerIPs))
-		for _, ip := range oldStatus.DNSServerIPs {
-			stringIPs = append(stringIPs, ip.String())
-		}
-		if len(stringIPs) == 0 {
-			return nil
-		}
-		logrus.Debugf("Removing DNS Servers %v from resolv.conf", stringIPs)
-		if err := c.removeNameserver(stringIPs); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-// ConnectNetwork connects a container to a given network
-func (c *Container) NetworkConnect(nameOrID, netName string, netOpts types.PerNetworkOptions) error {
-	// only the bridge mode supports cni networks
-	if err := isBridgeNetMode(c.config.NetMode); err != nil {
-		return err
-	}
-
-	c.lock.Lock()
-	defer c.lock.Unlock()
-
-	networks, err := c.networks()
-	if err != nil {
-		return err
-	}
-
-	// check if network exists and if the input is a ID we get the name
-	// CNI only uses names so it is important that we only use the name
-	netName, err = c.runtime.normalizeNetworkName(netName)
-	if err != nil {
-		return err
-	}
-
-	if err := c.syncContainer(); err != nil {
-		return err
-	}
-
-	// get network status before we connect
-	networkStatus := c.getNetworkStatus()
-
-	// always add the short id as alias for docker compat
-	netOpts.Aliases = append(netOpts.Aliases, c.config.ID[:12])
-
-	if netOpts.InterfaceName == "" {
-		netOpts.InterfaceName = getFreeInterfaceName(networks)
-		if netOpts.InterfaceName == "" {
-			return errors.New("could not find free network interface name")
-		}
-	}
-
-	if err := c.runtime.state.NetworkConnect(c, netName, netOpts); err != nil {
-		return err
-	}
-	c.newNetworkEvent(events.NetworkConnect, netName)
-	if !c.ensureState(define.ContainerStateRunning, define.ContainerStateCreated) {
-		return nil
-	}
-
-	opts := types.NetworkOptions{
-		ContainerID:   c.config.ID,
-		ContainerName: getCNIPodName(c),
-	}
-	opts.PortMappings = c.convertPortMappings()
-	opts.Networks = map[string]types.PerNetworkOptions{
-		netName: netOpts,
-	}
-
-	/*
-		results, err := c.runtime.setUpNetwork(c.state.NetNS.Path(), opts)
-		if err != nil {
-			return err
-		}
-		if len(results) != 1 {
-			return errors.New("when adding aliases, results must be of length 1")
-		}
-	*/
-	var results map[string]types.StatusBlock
-
-	// update network status
-	if networkStatus == nil {
-		networkStatus = make(map[string]types.StatusBlock, 1)
-	}
-	networkStatus[netName] = results[netName]
-	c.state.NetworkStatus = networkStatus
-
-	err = c.save()
-	if err != nil {
-		return err
-	}
-
-	// The first network needs a port reload to set the correct child ip for the rootlessport process.
-	// Adding a second network does not require a port reload because the child ip is still valid.
-
-	ipv6, err := c.checkForIPv6(networkStatus)
-	if err != nil {
-		return err
-	}
-
-	// Update resolv.conf if required
-	stringIPs := make([]string, 0, len(results[netName].DNSServerIPs))
-	for _, ip := range results[netName].DNSServerIPs {
-		if (ip.To4() == nil) && !ipv6 {
-			continue
-		}
-		stringIPs = append(stringIPs, ip.String())
-	}
-	if len(stringIPs) == 0 {
-		return nil
-	}
-	logrus.Debugf("Adding DNS Servers %v to resolv.conf", stringIPs)
-	if err := c.addNameserver(stringIPs); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// get a free interface name for a new network
-// return an empty string if no free name was found
-func getFreeInterfaceName(networks map[string]types.PerNetworkOptions) string {
-	ifNames := make([]string, 0, len(networks))
-	for _, opts := range networks {
-		ifNames = append(ifNames, opts.InterfaceName)
-	}
-	for i := 0; i < 100000; i++ {
-		ifName := fmt.Sprintf("eth%d", i)
-		if !util.StringInSlice(ifName, ifNames) {
-			return ifName
-		}
-	}
-	return ""
-}
-
-// DisconnectContainerFromNetwork removes a container from its CNI network
-func (r *Runtime) DisconnectContainerFromNetwork(nameOrID, netName string, force bool) error {
-	ctr, err := r.LookupContainer(nameOrID)
-	if err != nil {
-		return err
-	}
-	return ctr.NetworkDisconnect(nameOrID, netName, force)
-}
-
-// ConnectContainerToNetwork connects a container to a CNI network
-func (r *Runtime) ConnectContainerToNetwork(nameOrID, netName string, netOpts types.PerNetworkOptions) error {
-	ctr, err := r.LookupContainer(nameOrID)
-	if err != nil {
-		return err
-	}
-	return ctr.NetworkConnect(nameOrID, netName, netOpts)
-}
-
-// normalizeNetworkName takes a network name, a partial or a full network ID and returns the network name.
-// If the network is not found a errors is returned.
-func (r *Runtime) normalizeNetworkName(nameOrID string) (string, error) {
-	net, err := r.network.NetworkInspect(nameOrID)
-	if err != nil {
-		return "", err
-	}
-	return net.Name, nil
+func (c *Container) reloadRootlessRLKPortMapping() error {
+	return errors.New("unsupported (*Container).reloadRootlessRLKPortMapping")
 }
 
 // ocicniPortsToNetTypesPorts convert the old port format to the new one
